+++ point-of-sale-sdk/src/main/java/com/squareup/sdk/pos/transaction/PosApiTransactionAdapterFactory.java
@@ -1,0 +1,14 @@
+package com.squareup.sdk.pos.transaction;
+
+import com.google.gson.TypeAdapterFactory;
+import com.ryanharter.auto.value.gson.GsonTypeAdapterFactory;
+
+@GsonTypeAdapterFactory //
+public abstract class PosApiTransactionAdapterFactory
+    implements TypeAdapterFactory {
+
+  public static TypeAdapterFactory create() {
+    return new AutoValueGson_PosApiTransactionAdapterFactory();
+  }
+}
+