+++ sample-bikeshop/src/main/java/com/example/owensbikes/Option.java
@@ -40,4 +40,3 @@ class Option implements Serializable {
     return price_money.amount;
   }
 }
-